+++ tango/integrations/transformers/__init__.py
@@ -186,10 +186,19 @@ library under the corresponding class from the `torch <torch.html>`_ integration
             transformers::DefaultDataCollator
 """
 
-__all__ = ["RunGeneration", "RunGenerationDataset", "Tokenizer", "Config"]
+__all__ = [
+    "RunGeneration",
+    "RunGenerationDataset",
+    "Tokenizer",
+    "Config",
+    "FinetuneWrapper",
+    "FinetuneStep",
+    "TokenizeText2TextData",
+]
 
 from .config import Config
 from .data import *  # noqa: F403
+from .finetune import FinetuneStep, FinetuneWrapper, TokenizeText2TextData
 from .model import *  # noqa: F403
 from .optim import *  # noqa: F403
 from .run_generation import RunGeneration, RunGenerationDataset